Revert "Model, ModelManager: refactored to use drink rather than person or addressbook"
This reverts commit 7c3bc8afccaeeb3a69b0564c2d6acb5345540c86.
