+++ src/main/java/seedu/address/model/Model.java
@@ -3,52 +3,76 @@ package seedu.address.model;
 import java.util.function.Predicate;
 
 import javafx.collections.ObservableList;
-import seedu.address.model.drink.Drink;
+import seedu.address.model.person.Person;
 
 /**
  * The API of the Model component.
  */
 public interface Model {
     /** {@code Predicate} that always evaluate to true */
-    Predicate<Drink> PREDICATE_SHOW_ALL_DRINKS = unused -> true;
+    Predicate<Person> PREDICATE_SHOW_ALL_PERSONS = unused -> true;
 
     /** Clears existing backing model and replaces with the provided new data. */
-    void resetData(ReadOnlyInventoryList newData);
+    void resetData(ReadOnlyAddressBook newData);
 
-    /** Returns the InventoryList */
-    ReadOnlyInventoryList getInventoryList();
+    /** Returns the AddressBook */
+    ReadOnlyAddressBook getAddressBook();
 
     /**
-     * Returns true if a drink with the same identity as {@code drink} exists in the address book.
+     * Returns true if a person with the same identity as {@code person} exists in the address book.
      */
-    boolean hasDrink(Drink drink);
+    boolean hasPerson(Person person);
 
     /**
-     * Deletes the given drink.
-     * The drink must exist in the inventory list.
+     * Deletes the given person.
+     * The person must exist in the address book.
      */
-    void deleteDrink(Drink target);
+    void deletePerson(Person target);
 
     /**
-     * Adds the given drink.
-     * {@code drink} must not already exist in the address book.
+     * Adds the given person.
+     * {@code person} must not already exist in the address book.
      */
-    void addDrink(Drink drink);
+    void addPerson(Person person);
 
     /**
-     * Replaces the given drink {@code target} with {@code editedDrink}.
-     * {@code target} must exist in the inventory list.
-     * The drink identity of {@code editedDrink} must not be the same as another existing drink in the inventory list.
+     * Replaces the given person {@code target} with {@code editedPerson}.
+     * {@code target} must exist in the address book.
+     * The person identity of {@code editedPerson} must not be the same as another existing person in the address book.
      */
-    // void updateDrink(Drink target, Drink editedDrink);
+    void updatePerson(Person target, Person editedPerson);
 
-    /** Returns an unmodifiable view of the filtered inventory list */
-    ObservableList<Drink> getFilteredDrinkList();
+    /** Returns an unmodifiable view of the filtered person list */
+    ObservableList<Person> getFilteredPersonList();
 
     /**
-     * Updates the filter of the filtered drink list to filter by the given {@code predicate}.
+     * Updates the filter of the filtered person list to filter by the given {@code predicate}.
      * @throws NullPointerException if {@code predicate} is null.
      */
-    void updateFilteredDrinkList(Predicate<Drink> predicate);
+    void updateFilteredPersonList(Predicate<Person> predicate);
 
+    /**
+     * Returns true if the model has previous address book states to restore.
+     */
+    boolean canUndoAddressBook();
+
+    /**
+     * Returns true if the model has undone address book states to restore.
+     */
+    boolean canRedoAddressBook();
+
+    /**
+     * Restores the model's address book to its previous state.
+     */
+    void undoAddressBook();
+
+    /**
+     * Restores the model's address book to its previously undone state.
+     */
+    void redoAddressBook();
+
+    /**
+     * Saves the current address book state for undo/redo.
+     */
+    void commitAddressBook();
 }
